+++ corvis/src/tracker/rc_tracker.cpp
@@ -734,17 +734,17 @@ size_t rc_getCalibration(rc_Tracker *tracker, const char **buffer)
 
     size_t size = std::min(tracker->sfm.accelerometers.size(), tracker->sfm.gyroscopes.size());
     cal.imus.resize(size);
-    for (int id = 0; id < size; id++) {
+    for (size_t id = 0; id < size; id++) {
         rc_describeGyroscope(tracker, id, &cal.imus[id].extrinsics, &cal.imus[id].intrinsics.gyroscope);
         rc_describeAccelerometer(tracker, id, &cal.imus[id].extrinsics, &cal.imus[id].intrinsics.accelerometer);
     }
 
     cal.cameras.resize(tracker->sfm.cameras.size());
-    for (int id = 0; id < tracker->sfm.cameras.size(); id++)
+    for (size_t id = 0; id < tracker->sfm.cameras.size(); id++)
         rc_describeCamera(tracker, id, rc_FORMAT_GRAY8, &cal.cameras[id].extrinsics, &cal.cameras[id].intrinsics);
 
     cal.depths.resize(tracker->sfm.depths.size());
-    for (int id = 0; id < tracker->sfm.depths.size(); id++)
+    for (size_t id = 0; id < tracker->sfm.depths.size(); id++)
         rc_describeCamera(tracker, id, rc_FORMAT_DEPTH16, &cal.depths[id].extrinsics, &cal.depths[id].intrinsics);
 
     std::string json;
